moved home and info into is own folder named pages and react router functioning.
