+++ src/components/Filter/Filter.tsx
@@ -1,0 +1,11 @@
+// import { Button } from "@/components/ui/button";
+
+export default function Filter() {
+  return (
+    <>
+      <div>
+        <h1>Hello from Filter</h1>
+      </div>
+    </>
+  );
+}
