+++ src/pages/InfoPage/InfoPage.tsx
@@ -1,0 +1,11 @@
+// import { Button } from "@/components/ui/button";
+
+export default function InfoPage() {
+  return (
+    <>
+      <div>
+        <h1>Hello from InfoPage</h1>
+      </div>
+    </>
+  );
+}
